+++ persistence/sqlite3.go
@@ -234,7 +234,7 @@ func (db *sqlite3Database) QueryTorrents(
 	orderBy OrderingCriteria,
 	ascending bool,
 	limit uint64,
-	lastOrderedValue *float64,
+	lastOrderedValue *uint64,
 	lastID *uint64,
 ) ([]TorrentMetadata, error) {
 	if query == "" && orderBy == ByRelevance {
